requirements.txt
numpy
email-validator
psycopg2
bcrypt
PyJWT

